Don't call stls in popStore when tls is disabled
Connecting to a POP3 mailserver without TLS fails. In the popStore.js at line 32 the stls function is called even if enabletls is false. This throws an stls error which causes the client to quit.

This change calls stls only when tls is enabled. When tls is disabled, the client immediatly authenticates.

PS Great library!